F4 - Scan subscribed clans for wars
First, removed the old code that managed user-input of new wars. This took out just about the whole war class, but going to keep the file in just a bit longer.

Second, added a scheduled task that runs every eight hours looking for new clan wars on all clans that are subscribed. Might switch this to four or twelve hours, depending on how things go.

some more work needs to be done to get the new WarRunners to send actual pings to end users.