F10 Add a connection pool for DB Connections
F10 Add a connection pool for DB Connections
Having one connection is great and all, but it may not be thread safe as it is right now. Fix that by implementing a pool structure.

Additionally, database connections go stale after about 8 hours. Hopefully this will automatically fix that. but I may need to do checks on the connections before using them to completely fix this half of the problem.