F23: Make subscriptions server-specific
Database change to add an extra field that allows subscriptions to be server-specific. This is the cleaner way to do subscriptions to make sure users in multiple discords have as much control as possible over their subscriptions. It will also allow us to refactor the WarChecker/Runner code in the near future to run more efficiently and look cleaner.